% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{detect_outliers_fixed}
\alias{detect_outliers_fixed}
\title{Fixed residual outlier rule}
\usage{
detect_outliers_fixed(residual, cutoff = 6)
}
\arguments{
\item{residual}{Numeric residuals.}

\item{cutoff}{Absolute residual cutoff.}
}
\value{
Logical vector where \code{TRUE} indicates flagged observations.
}
\description{
Fixed residual outlier rule
}
