% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/features.R
\name{eml_nmea}
\alias{eml_nmea}
\title{Estimated morning level from NMEA parameters}
\usage{
eml_nmea(psi)
}
\arguments{
\item{psi}{Matrix or data frame with columns \code{mu}, \code{s}, \code{c1}, \code{c0}, \code{alpha}.}
}
\value{
Numeric vector of estimated morning level values.
}
\description{
Estimated morning level from NMEA parameters
}
