% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_nmea.R
\name{predict_nmea}
\alias{predict_nmea}
\title{Predict NMEA curves from fitted or supplied parameters}
\usage{
predict_nmea(object, time_grid)
}
\arguments{
\item{object}{An \code{nmea_fit} object returned by \code{\link[=fit_nmea]{fit_nmea()}} or a parameter
matrix with columns \code{mu}, \code{s}, \code{c1}, \code{c0}, and \code{alpha}.}

\item{time_grid}{Numeric vector of times since waking.}
}
\value{
A numeric prediction matrix.
}
\description{
Predict NMEA curves from fitted or supplied parameters
}
