% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_parameters.R
\name{simulate_latent_curves}
\alias{simulate_latent_curves}
\title{Simulate latent NMEA cortisol curves}
\usage{
simulate_latent_curves(psi, time_grid = seq(0, 18, by = 0.1))
}
\arguments{
\item{psi}{Parameter matrix with columns \code{mu}, \code{s}, \code{c1}, \code{c0}, \code{alpha}.}

\item{time_grid}{Numeric time grid.}
}
\value{
A list containing parameters, time grid, and curve matrix.
}
\description{
Simulate latent NMEA cortisol curves
}
