% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_metrics.R
\name{summarize_curve_metrics}
\alias{summarize_curve_metrics}
\title{Summarize subject-level benchmark metrics}
\usage{
summarize_curve_metrics(metric_df)
}
\arguments{
\item{metric_df}{Output from \code{curve_benchmark_metrics()}.}
}
\value{
One-row data frame of aggregate metrics.
}
\description{
Summarize subject-level benchmark metrics
}
